Add a more permissive part validation regex
This adds a more permissive CPE URI validation regex. It permits literal
`\`, `!` and `+` characters in CPE URIs. Normally, these need to use
`%`-escapes, but this provides a compromise to prevent automated tool
failures gating builds because of almost-correct CPEs. This deliberately
does not include the full list of characters in section 5.4 of the CPE
2.2 specification, but rather focuses on the most common invalid
characters we've seen in real world data.